+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # v1.4.4 (Unreleased)
+### Bug fixes
+* Updated the `pingfederate_authentication_policies` resource to set no default for custom attribute source `filter_fields.value` attributes, and to validate that the configured filter field value string has length at least 1. This will prevent inconsistent result errors when using custom attribute sources in the authentication policies. Related to a known terraform-plugin-framework bug with defaults in nested sets: [#867](https://github.com/hashicorp/terraform-plugin-framework/issues/867). ([#484](https://github.com/pingidentity/terraform-provider-pingfederate/pull/484))
+
 ### Notes
 * bump Go 1.23.5 => 1.24.1 ([#478](https://github.com/pingidentity/terraform-provider-pingfederate/pull/478))